src/requirements.txt
Beaker
Markdown
dulwich
Mercurial
bottle
bottle-werkzeug
bottle-cork
Werkzeug
peewee
